+++ Sistema.Entidades/Almacen/Sucursal.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations.Schema;
 
 namespace Sistema.Entidades.Almacen
@@ -5,7 +6,7 @@ namespace Sistema.Entidades.Almacen
     public class Sucursal
     {
         [Column("suc_id")]
-        public int SucId { get; set; }
+        public Int16 SucId { get; set; }
         [Column("suc_nombre")]
         public string SucNombre { get; set; }
     }
