+++ Sistema.Web/Models/Almacen/AlmacenProducto/ActualizarViewModel.cs
@@ -7,6 +7,6 @@ namespace Sistema.Web.Models.Almacen.AlmacenProducto
         [Key]
         public int AlmId { get; set; }
         public int ProId { get; set; }
-        public float AlpStockActual { get; set; }
+        public double AlpStockActual { get; set; }
     }
 }
