+++ Sistema.Web/Models/Almacen/AlmacenProducto/AlmacenProductoViewModel.cs
@@ -4,7 +4,7 @@ namespace Sistema.Web.Models.Almacen.AlmacenProducto
     {
         public int AlmId { get; set; }
         public int ProId { get; set; }
-        public float AlpStockActual { get; set; }
+        public double AlpStockActual { get; set; }
         public string almacen { get; set; }
         public string producto { get; set; }
     }
