+++ Sistema.Web/Models/Almacen/Sucursal/CrearViewModel.cs
@@ -1,3 +1,4 @@
+using System;
 using System.ComponentModel.DataAnnotations;
 
 namespace Sistema.Web.Models.Almacen.Sucursal
@@ -5,7 +6,7 @@ namespace Sistema.Web.Models.Almacen.Sucursal
     public class CrearViewModel
     {
         [Key]
-        public int SucId { get; set; }
+        public Int16 SucId { get; set; }
         public string SucNombre { get; set; }
     }
 }
